Refactoring: Move node logic in separate file.
